refactor: split OpenCode config hook concerns
